Refactored Date formatting.. as converting String and Double to Date first and then Date back to String in proper format..
