+++ Weather/Extensions/Double.swift
@@ -2,20 +2,12 @@
 //  Double.swift
 //  Weather
 //
-//  Created by Kautsya Kanu on 22/02/21.
+//  Created by Kautsya Kanu on 23/02/21.
 //
 
 import Foundation
 extension Double {
-    var localTime: String {
-        let date = Date(timeIntervalSince1970: self)
-        let dateFormatter = DateFormatter()
-        dateFormatter.timeStyle = .medium
-        dateFormatter.dateStyle = .medium
-        dateFormatter.timeZone = .current
-        dateFormatter.dateFormat = "h:mm a"
-        dateFormatter.amSymbol = "AM"
-        dateFormatter.pmSymbol = "PM"
-        return dateFormatter.string(from: date)
+    var date: Date? {
+        return Date(timeIntervalSince1970: self)
     }
 }
